% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBezierCurve.R
\name{addBezierCurve}
\alias{addBezierCurve}
\title{Add Bezier Curve Lines
(https://ryancatalani.medium.com/creating-consistently-curved-lines-on-leaflet-b59bc03fa9dc)}
\usage{
addBezierCurve(map, options = list(lng1, lat1, lng2, lat2, color, weight))
}
\arguments{
\item{map}{map}

\item{options}{paramétrage de la courbe
(liste à 6 élements : lng1, lat1, lng2, lat2, color, weight)}
}
\description{
Add Bezier Curve Lines
(https://ryancatalani.medium.com/creating-consistently-curved-lines-on-leaflet-b59bc03fa9dc)
}
\examples{
library(leaflet)
m <- leaflet() \%>\%
 addTiles() \%>\%
 addBezierCurve(options =
                  list(
                    lng1 = 2.3488,
                    lat1 = 48.8534,
                    lng2 = -0.57918,
                    lat2 = 44.837789,
                    color = "blue",
                    weight = 10
                  )
 )
}
